requirements.txt
ollama>=0.1.15
python-dateutil>=2.8.2
pathlib>=1.0.1
